Reconfigure the fs test(2). To enable unit test of cross-compiled enclave, the OS-specified tests must be co nfigurable at host side. A host library to remove non-empty directory is introduced, in case this function can be reused by other tests.
