+++ tests/syscall/CMakeLists.txt
@@ -7,6 +7,7 @@ add_subdirectory(fs)
 add_subdirectory(hostfs)
 add_subdirectory(resolver)
 add_subdirectory(socket)
+add_subdirectory(tool)
 if(UNIX)
 add_subdirectory(datagram)
 add_subdirectory(epoll)
