+++ tests/syscall/fs/host/host_linux.c
@@ -1,45 +1,17 @@
 // Copyright (c) Open Enclave SDK contributors.
 // Licensed under the MIT License.
 
-#define _XOPEN_SOURCE 500
-
-#include <ftw.h>
 #include <openenclave/host.h>
 #include <openenclave/internal/syscall/host.h>
 #include <openenclave/internal/tests.h>
 #include <stdio.h>
 #include <sys/stat.h>
 #include <sys/types.h>
-#include <unistd.h>
 #include "fs_u.h"
 
 #define SKIP_RETURN_CODE 2
 
-static int do_delete(
-    const char* path,
-    const struct stat* sb,
-    int tflag,
-    struct FTW* ftwbuf)
-{
-    switch (tflag)
-    {
-        case FTW_D:
-        case FTW_DNR:
-        case FTW_DP:
-            rmdir(path);
-            break;
-        default:
-            unlink(path);
-            break;
-    }
-
-    return (0);
-}
-int recursive_rmdir(const char* path)
-{
-    // DFS depth is set to 64.
-    return nftw(path, do_delete, 64, FTW_DEPTH);
-}
+int recursive_rmdir(const char* path);
 
 int main(int argc, const char* argv[])
 {
