+++ tests/syscall/fs/host/host_win32.c
@@ -1,49 +1,16 @@
 // Copyright (c) Open Enclave SDK contributors.
 // Licensed under the MIT License.
 
-#pragma comment(lib, "Shell32.lib")
-// windows.h first to avoid conflicts.
 #include <windows.h>
 
 #include <openenclave/host.h>
 #include <openenclave/internal/syscall/host.h>
 #include <openenclave/internal/tests.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <sys/stat.h>
-#include <sys/types.h>
 #include "fs_u.h"
 
 #define SKIP_RETURN_CODE 2
 
-int rmdir(wchar_t* path)
-{
-    int ret = -1;
-    wchar_t* doublenullpath = NULL;
-    int len = (int)wcslen(path);
-    SHFILEOPSTRUCTW opt;
-
-    doublenullpath = malloc((len + 2) * sizeof(wchar_t));
-    if (!doublenullpath)
-    {
-        goto done;
-    }
-    memcpy(doublenullpath, path, len * sizeof(wchar_t));
-    doublenullpath[len] = doublenullpath[len + 1] = L'\0';
-
-    memset(&opt, 0, sizeof(SHFILEOPSTRUCTW));
-    opt.pFrom = doublenullpath;
-    opt.hwnd = NULL;
-    opt.wFunc = FO_DELETE;
-    opt.fFlags = FOF_SILENT | FOF_NOERRORUI | FOF_ALLOWUNDO |
-                 FOF_NOCONFIRMMKDIR | FOF_NOCONFIRMATION;
-
-    ret = SHFileOperationW(&opt);
-
-done:
-    free(doublenullpath);
-    return ret;
-}
+int recursive_rmdir(wchar_t* path);
 
 int wmain(int argc, wchar_t* argv[])
 {
@@ -86,7 +53,7 @@ int wmain(int argc, wchar_t* argv[])
     // Windows does not support umask.
     // Please set up the right permission to the parent directory.
 
-    rmdir(argv[3]);
+    recursive_rmdir(argv[3]);
 
     r = oe_create_fs_enclave(enclave_path, type, flags, NULL, 0, &enclave);
     OE_TEST(r == OE_OK);
