+++ tests/syscall/fs/windows/fs.edl
@@ -1,0 +1,11 @@
+// Copyright (c) Open Enclave SDK contributors.
+// Licensed under the MIT License.
+
+enclave {
+
+    trusted {
+        public void test_fs(
+            [string, in] const char* src_dir,
+            [string, in] const char* tmp_dir);
+    };
+};
