+++ tests/syscall/hostfs/CMakeLists.txt
@@ -9,12 +9,4 @@ endif()
 
 set(TMP_DIR "${CMAKE_CURRENT_BINARY_DIR}/tmp")
 
-add_test(tests/hostfs1 cmake -E remove_directory "${TMP_DIR}")
-
 add_enclave_test(tests/hostfs hostfs_host hostfs_enc "${TMP_DIR}")
-
-# Ensuring removing the tmp directory before the second test run.
-if (LVI_MITIGATION MATCHES ControlFlow AND NOT LVI_MITIGATION_SKIP_TESTS)
-    add_test(tests/hostfs1-lvi-cfg cmake -E remove_directory "${TMP_DIR}")
-    set_tests_properties(tests/hostfs-lvi-cfg PROPERTIES DEPENDS tests/hostfs1-lvi-cfg)
-endif()
