+++ tests/syscall/hostfs/host/CMakeLists.txt
@@ -15,3 +15,4 @@ add_executable(hostfs_host host.c test_hostfs_u.c)
 target_include_directories(hostfs_host PRIVATE ${CMAKE_CURRENT_BINARY_DIR})
 
 target_link_libraries(hostfs_host oehostapp)
+target_link_libraries(hostfs_host rmdir)
